backend/requirements.txt
fastapi
uvicorn
python-dotenv
passlib[bcrypt]
pydantic
sqlalchemy
python-jose
langchain
langchain-google-genai
faiss-cpu
langchain-community
psycopg2-binary